Source Code/Linux/requirement-linux.txt
PyQt5
pyinstaller
keyboard
schedule
pillow
pyalsaaudio